SSE-2732: Disambiguate SAM and Docker paths
This action builds two things (SAM app and a Docker image) and therefore needs two build directories, defaulting to the current dir, as we cannot assume they're the same.

We already have a docker build path so replace the "working dir" with the SAM base dir to avoid confusion. The singular "working directory" is confusing and misunderstood, which causes breaking changes to the action being made.

The working-directory input is retained for backwards compatibility and is used as the SAM build dir if the new param is not set.

For Docker, allow dockerfile and docker path to be different since the dockerfile doesn't necessarily always reside in the docker path so the filename cannot be prefixed with the build path by default - use the docker build args accordingly.

Allow the SAM template file to be different from the base build dir - the path from which SAM resources are resolved is not always the same as the template file, which is the default, therefore the --base-dir param must be used.

Don't cd into any directory as the current working dir is set by the calling workflow, and it's unexpected behaviour to cd away from it.

Set default working dirs in the script
The "." default value doesn't get carried to the script env vars. Handle default value of null in the script itself.